Add proper handling for a datastore command and subcommands
